add wokcommands and move ready event to index.ts
