+++ package.json
@@ -24,7 +24,8 @@
     "dayjs": "^1.10.8",
     "discord.js": "^13.6.0",
     "dotenv": "^14.2.0",
-    "google-spreadsheet": "^3.2.0"
+    "google-spreadsheet": "^3.2.0",
+    "wokcommands": "^1.5.3"
   },
   "devDependencies": {
     "@types/google-spreadsheet": "^3.1.5",
